// src/app/common/models/zoneInfo.ts
export interface iZoneInfo {
    zone_full_name: string;
    iso2: string;
    zone_short_name: string;
    continent: string;
    sub_region: string;
    population: number;
    capitol: string;
    major_geography: string;
    predominant_language: string;
    driving_side: string;
    peak_tourist_season: string;
    best_time_to_dive: string;
    bad_time_to_go: string;
    bodies_of_water: string;
    zone_description: string;
}